userService/pyproject.toml
[tool.poetry]
name = "userservice"
version = "0.1.0"
description = ""
authors = ["Saim khalid <saimkhalid6220@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlmodel = "^0.0.19"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.2.1"}
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
email-validator = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
